// .repos/beep-effect/packages/common/utils/src/sanitize-html/types.ts
export interface Attributes {
  [attr: string]: string;
}

/**
 * Result of a tag transformation.
 *
 * @since 0.1.0
 * @category types
 */
export interface TransformedTag {
  readonly tagName: string;
  readonly attribs: Attributes;
  readonly text?: undefined | string;
}

/**
 * Tag transformation function type.
 *
 * @since 0.1.0
 * @category types
 */
export type Transformer = (tagName: string, attribs: Attributes) => TransformedTag;